server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
aiosqlite==0.21.0
python-jose[cryptography]==3.3.0
google-auth==2.40.0
pydantic-settings==2.9.1
python-multipart==0.0.20
requests>=2.32
